refactor: split lib.rs into per-concern modules
Pure code motion off upstream main, no behaviour change, as the base for
the typed rewrite.
