+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "cjseq"
-version = "0.4.1"
-description = "Create+process+modify+convert CityJSONSeq"
+name = "cjseq2"
+version = "0.2.0-alpha.0"
+description = "Typed fork of cjseq (create+process+modify+convert CityJSONSeq) pending upstream merge"
 authors = ["Hugo Ledoux <h.ledoux@tudelft.nl>"]
 edition = "2021"
 license = "MIT"
@@ -10,6 +10,7 @@ homepage = "https://github.com/cityjson/cjseq"
 repository = "https://github.com/cityjson/cjseq"
 
 [lib]
+name = "cjseq"
 crate-type = ["cdylib", "rlib"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
