+++ src/appearance.rs
@@ -1,0 +1,92 @@
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct Material {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub values: Option<Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub value: Option<usize>,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct Texture {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub values: Option<Value>,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct Appearance {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub materials: Option<Vec<Value>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub textures: Option<Vec<Value>>,
+    #[serde(rename = "vertices-texture")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub vertices_texture: Option<Vec<Vec<f64>>>,
+    #[serde(rename = "default-theme-texture")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub default_theme_texture: Option<String>,
+    #[serde(rename = "default-theme-material")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub default_theme_material: Option<String>,
+}
+impl Appearance {
+    pub(crate) fn new() -> Self {
+        Appearance {
+            materials: None,
+            textures: None,
+            vertices_texture: None,
+            default_theme_texture: None,
+            default_theme_material: None,
+        }
+    }
+    pub(crate) fn add_material(&mut self, jm: Value) -> usize {
+        let re = match &mut self.materials {
+            Some(x) => match x.iter().position(|e| *e == jm) {
+                Some(y) => y,
+                None => {
+                    x.push(jm);
+                    x.len() - 1
+                }
+            },
+            None => {
+                let mut ls: Vec<Value> = Vec::new();
+                ls.push(jm);
+                self.materials = Some(ls);
+                0
+            }
+        };
+        re
+    }
+    pub(crate) fn add_texture(&mut self, jm: Value) -> usize {
+        let re = match &mut self.textures {
+            Some(x) => match x.iter().position(|e| *e == jm) {
+                Some(y) => y,
+                None => {
+                    x.push(jm);
+                    x.len() - 1
+                }
+            },
+            None => {
+                let mut ls: Vec<Value> = Vec::new();
+                ls.push(jm);
+                self.textures = Some(ls);
+                0
+            }
+        };
+        re
+    }
+    pub(crate) fn add_vertices_texture(&mut self, mut vs: Vec<Vec<f64>>) {
+        match &mut self.vertices_texture {
+            Some(x) => {
+                x.append(&mut vs);
+            }
+            None => {
+                let mut ls: Vec<Vec<f64>> = Vec::new();
+                ls.append(&mut vs);
+                self.vertices_texture = Some(ls);
+            }
+        };
+    }
+}
