+++ src/city_object.rs
@@ -1,0 +1,52 @@
+use crate::geometry::Geometry;
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct CityObject {
+    #[serde(rename = "type")]
+    pub thetype: String,
+    #[serde(rename = "geographicalExtent")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub geographical_extent: Option<Vec<f64>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub attributes: Option<Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub geometry: Option<Vec<Geometry>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub children: Option<Vec<String>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub parents: Option<Vec<String>>,
+    #[serde(flatten)]
+    other: serde_json::Value,
+}
+
+impl CityObject {
+    pub fn get_type(&self) -> String {
+        self.thetype.clone()
+    }
+    pub(crate) fn is_toplevel(&self) -> bool {
+        match &self.parents {
+            Some(x) => {
+                if x.is_empty() {
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+            None => return true,
+        }
+    }
+    pub(crate) fn get_children_keys(&self) -> Vec<String> {
+        let mut re: Vec<String> = Vec::new();
+        match &self.children {
+            Some(x) => {
+                for each in x {
+                    re.push(each.to_string());
+                }
+            }
+            None => (),
+        }
+        re
+    }
+}
