+++ src/metadata.rs
@@ -1,0 +1,150 @@
+use serde::{Deserialize, Serialize};
+
+const DEFAULT_CRS_BASE_URL: &str = "https://www.opengis.net/def/crs";
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct Transform {
+    pub scale: Vec<f64>,
+    pub translate: Vec<f64>,
+}
+impl Transform {
+    pub(crate) fn new() -> Self {
+        Transform {
+            scale: vec![1.0, 1.0, 1.0],
+            translate: vec![0., 0., 0.],
+        }
+    }
+}
+
+pub type GeographicalExtent = [f64; 6];
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct Address {
+    #[serde(rename = "thoroughfareNumber")]
+    pub thoroughfare_number: i64,
+    #[serde(rename = "thoroughfareName")]
+    pub thoroughfare_name: String,
+    pub locality: String,
+    #[serde(rename = "postalCode")]
+    pub postal_code: String,
+    pub country: String,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct PointOfContact {
+    #[serde(rename = "contactName")]
+    pub contact_name: String,
+    #[serde(rename = "contactType")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub contact_type: Option<String>,
+    #[serde(rename = "role")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub role: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub phone: Option<String>,
+    #[serde(rename = "emailAddress")]
+    pub email_address: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub website: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub address: Option<Address>,
+}
+
+/// A reference system following the OGC Name Type Specification.
+///
+/// The format follows: `http://www.opengis.net/def/crs/{authority}/{version}/{code}`
+/// where:
+/// - `{authority}` designates the authority responsible for the definition of this CRS
+///   (usually "EPSG" or "OGC")
+/// - `{version}` designates the specific version of the CRS
+///   (use "0" if there is no version)
+/// - `{code}` is the identifier for the specific coordinate reference system
+#[derive(Debug, Clone)]
+pub struct ReferenceSystem {
+    pub base_url: String,
+    pub authority: String,
+    pub version: String,
+    pub code: String,
+}
+
+impl ReferenceSystem {
+    pub fn new(base_url: Option<String>, authority: String, version: String, code: String) -> Self {
+        let base_url = base_url.unwrap_or(DEFAULT_CRS_BASE_URL.to_string());
+        ReferenceSystem {
+            base_url,
+            authority,
+            version,
+            code,
+        }
+    }
+
+    pub fn to_url(&self) -> String {
+        format!(
+            "{}/{}/{}/{}",
+            self.base_url, self.authority, self.version, self.code
+        )
+    }
+
+    // OGC Name Type Specification:
+    // http://www.opengis.net/def/crs/{authority}/{version}/{code}
+    // where {authority} designates the authority responsible for the definition of this CRS (usually "EPSG" or "OGC"), and where {version} designates the specific version of the CRS ("0" (zero) is used if there is no version).
+    pub fn from_url(url: &str) -> Result<Self, &'static str> {
+        if !url.contains("//www.opengis.net/def/crs") {
+            return Err("Invalid reference system URL");
+        }
+
+        let i = url.find("crs").unwrap();
+        let s = &url[i + 4..];
+
+        let parts: Vec<&str> = s.split("/").collect();
+        if parts.len() != 3 {
+            return Err("Invalid reference system URL");
+        }
+
+        Ok(ReferenceSystem {
+            base_url: url[..i + 3].to_string(),
+            authority: parts[0].to_string(),
+            version: parts[1].to_string(),
+            code: parts[2].to_string(),
+        })
+    }
+}
+
+impl Serialize for ReferenceSystem {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        self.to_url().serialize(serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for ReferenceSystem {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let url = String::deserialize(deserializer)?;
+        ReferenceSystem::from_url(&url).map_err(serde::de::Error::custom)
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct Metadata {
+    #[serde(rename = "geographicalExtent")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub geographical_extent: Option<GeographicalExtent>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub identifier: Option<String>,
+    #[serde(rename = "pointOfContact")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub point_of_contact: Option<PointOfContact>,
+    #[serde(rename = "referenceDate")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub reference_date: Option<String>,
+    #[serde(rename = "referenceSystem")]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub reference_system: Option<ReferenceSystem>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub title: Option<String>,
+}
